EDSC-4055: Testing with an alternative placement.
