+++ tests/e2e/collection_list/collection_list.spec.js
@@ -34,4 +34,26 @@ test.describe('Collection List Behavior', () => {
       (await page.getByTestId('collection-results-table__item').all()).length
     ).toEqual(1)
   })
+
+  test('Search page load time is less than 1 second', async ({ page }) => {
+    const requestFinishedPromise = page.waitForEvent('requestfinished')
+    const request = await requestFinishedPromise
+
+    expect(request.timing().responseEnd < 5000).toBe(true)
+  })
+
+  test('Search page LCP start time is less than 7 second', async ({ page }) => {
+    const LCP = await page.evaluate(() => new Promise((resolve) => {
+      new PerformanceObserver((list) => {
+        const entries = list.getEntries()
+        const lcp = entries.at(-1)
+        resolve(lcp.startTime)
+      }).observe({
+        type: 'largest-contentful-paint',
+        buffered: true
+      })
+    }))
+
+    expect(LCP < 10000).toBe(true)
+  })
 })
